+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CpxE: Added CLI for writing and reading values.
 
 ### Changed
+- Removed timeout and port from CPX-base, implemented timeout for CPX-AP
 - Reworked unit tests
 - Moved typecode ids to cpx_*_definitions.py
 - e1ci dicts are now dataclasses